# api/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "API para gerenciamento de criptoativos.",
      "version": "1.0.0",
      "title": "API de Criptoativos"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
      "/explorar_dados": {
        "get": {
          "summary": "Explorar dados de um ativo",
          "parameters": [
            {
              "name": "ativo",
              "in": "query",
              "description": "O ativo a ser explorado",
              "required": false,
              "type": "string",
              "default": "BTC-USD"
            }
          ],
          "responses": {
            "200": {
              "description": "Dados explorados com sucesso"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/treinar_modelo": {
        "post": {
          "summary": "Treinar modelo para um ativo",
          "parameters": [
            {
              "name": "ativo",
              "in": "body",
              "description": "Ativo para treinar o modelo",
              "required": false,
              "schema": {
                "type": "object",
                "properties": {
                  "ativo": {
                    "type": "string",
                    "default": "BTC-USD"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Modelo treinado com sucesso"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/testar_modelo": {
        "post": {
          "summary": "Testar modelo para um ativo",
          "parameters": [
            {
              "name": "ativo",
              "in": "body",
              "description": "Ativo para testar o modelo",
              "required": false,
              "schema": {
                "type": "object",
                "properties": {
                  "ativo": {
                    "type": "string",
                    "default": "BTC-USD"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Teste realizado com sucesso"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      },
      "/prever_valores": {
        "post": {
          "summary": "Prever valores para um ativo",
          "parameters": [
            {
              "name": "ativo",
              "in": "body",
              "description": "Ativo para prever os valores",
              "required": false,
              "schema": {
                "type": "object",
                "properties": {
                  "ativo": {
                    "type": "string",
                    "default": "BTC-USD"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Previsões geradas com sucesso"
            },
            "500": {
              "description": "Erro interno do servidor"
            }
          }
        }
      }
    }
  }
  